Benchmark of count from ordered_index and ranked_index
A benchmark is added as count_benchmark.cpp in the example
directory.

In the cases when values of an index are essentially unique, the
implementation in ranked_index is about 1/3 slower.

However, for highly non-unique indices (like the age of people),
the new implementation in ranked_index outperforms ordered_index.
Already for 1000 people of ages between 0 and 99, ranked_index is
about 50% faster, for 10 000 people it is about 8 times faster,
and for 100 000 people about 80 times faster.

For even more non-unique indices, acquiring only 10 distinct
values, already for 10 000 elements ranked_index is about 80 times
faster than ordered_index, and for 100 000 elements about 900 times
faster.