Index with 100 values with comparison predicate differentiating 10
Before, to benchmark count in situations of high non-uniqueness,
in count_benchmark.cpp one would change max_age=100 to max_age=10
in order to consider an index with only 10 distinct values.

Now, in addition to the old benchmark with max_age=100, we use an
index with max_age=100 distinct values together with a comparison
predicate that differentiates 10 groups of these values, which is
a very natural use case. An advantage of this change is that both
functions count(x) and count(x,comp) are now covered by the
benchmark.

In the benchmark using the custom comparison predicate, the gap in
performance between count in ranked_index and ordered_index is
slightly smaller than in the old benchmark with the default
comparison predicate and index with 10 values: the former is almost
100 times faster for 10 000 elements in the container (compared
with ca. 120 times with the default comparison predicate), and
almost 1000 times faster for 100 000 elements (compared with over
1100 times).